requirements.txt
requests==2.32.2
google-cloud-storage==2.4.0
pyarrow==17.0.0
psycopg2==2.9.4
datasets==2.21.0
pyyaml==6.0
numpy<2.0

transformers
evaluate
soundfile
